+++ spec/pre_spec.rb
@@ -45,4 +45,21 @@ RSpec.describe(PreTagBlock) do
     END_CONTENT
     expect(numbered_content).to eq(expected_content)
   end
+
+  it "highlights regex patterns" do
+    content = <<~END_CONTENT
+      Line 1
+        Line 2
+          Line 3
+          Line 4
+        Line 5
+      Line 6
+      Line 7
+        Line 8
+        Line 9
+      Line 10
+    END_CONTENT
+    highlighted = PreTagBlock.highlight(content, ".*2").split("\n")[1]
+    expect(highlighted).to eq("<span class='bg_yellow'>  Line 2</span>")
+  end
 end