+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.3
   * Documented the `data-lt-active="false"` attribute.
+  * Added the `dark` option, and [provided CSS](https://www.mslinn.com/blog/2020/10/03/jekyll-plugins.html#pre_css).
 
 ## 1.1.2 / 2022-04-05
   * Updated to `jekyll_plugin_logger` v2.1.0